Move TT-API and XML functions to utiltt.h/cpp and utilxml.h/cpp
Reduces size and dependencies on common.h
