+++ Client/qtTeamTalk/common.cpp
@@ -36,124 +36,6 @@
 extern QSettings* ttSettings;
 extern TTInstance* ttInst;
 
-bool userCanTx(int userid, StreamTypes stream_type, const Channel& chan)
-{
-    int i=0;
-    while(i<TT_TRANSMITUSERS_MAX && chan.transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX])
-    {
-        if(chan.transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX] == userid && (chan.transmitUsers[i][TT_TRANSMITUSERS_STREAMTYPE_INDEX] & stream_type))
-            return (chan.uChannelType & CHANNEL_CLASSROOM) == CHANNEL_CLASSROOM;
-        else i++;
-    }
-    return (chan.uChannelType & CHANNEL_CLASSROOM) == CHANNEL_DEFAULT;
-}
-
-bool userCanChanMessage(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_CHANNELMSG, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_CHANNELMSG, chan));
-}
-
-bool userCanVoiceTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_VOICE, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_VOICE, chan));
-}
-
-bool userCanVideoTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_VIDEOCAPTURE, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_VIDEOCAPTURE, chan));
-}
-
-bool userCanDesktopTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_DESKTOP, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_DESKTOP, chan));
-}
-
-bool userCanMediaFileTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_MEDIAFILE, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_MEDIAFILE, chan));
-}
-
-channels_t getSubChannels(int channelid, const channels_t& channels, bool recursive /*= false*/)
-{
-    channels_t subchannels;
-    for(auto ite = channels.begin(); ite != channels.end(); ++ite)
-    {
-        if(ite.value().nParentID == channelid)
-        {
-            subchannels[ite.key()] = ite.value();
-#if QT_VERSION < QT_VERSION_CHECK(6,0,0)
-            if(recursive)
-                subchannels.unite(getSubChannels(ite.value().nChannelID, channels, recursive));
-#else
-                subchannels.insert(getSubChannels(ite.value().nChannelID, channels, recursive));
-#endif
-        }
-    }
-    return subchannels;
-}
-
-channels_t getParentChannels(int channelid, const channels_t& channels)
-{
-    channels_t parents;
-    while(channels[channelid].nParentID>0)
-    {
-        parents[channels[channelid].nParentID] = channels[channelid];
-        channelid = channels[channelid].nParentID;
-    }
-    return parents;
-}
-
-users_t getChannelUsers(int channelid, const users_t& users, const channels_t& channels, bool recursive /* = false */)
-{
-    users_t result;
-    for(auto ite=users.begin();ite!=users.end();ite++)
-    {
-        if(ite.value().nChannelID == channelid)
-            result.insert(ite.key(), ite.value());
-    }
-
-    if (recursive)
-    {
-        channels_t subs = getSubChannels(channelid, channels, true);
-        for(auto i=subs.begin();i!=subs.end();++i)
-        {
-            for (auto& u : getChannelUsers(i.key(), users, channels))
-                result.insert(u.nUserID, u);
-        }
-    }
-    return result;
-}
-
-bool isFreeForAll(StreamTypes stream_type, const int transmitUsers[][2],
-                  int max_userids /*= TT_TRANSMITUSERS_MAX*/)
-{
-    int i=0;
-    while(i<max_userids && transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX] != 0)
-    {
-        if(transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX] == TT_CLASSROOM_FREEFORALL &&
-           (transmitUsers[i][TT_TRANSMITUSERS_STREAMTYPE_INDEX] & stream_type))
-            return true;
-        i++;
-    }
-    return false;
-}
-
-void setTransmitUsers(const QSet<int>& users, INT32* dest_array,
-                      INT32 max_elements)
-{
-    QSet<int>::const_iterator ite = users.begin();
-    for(int i=0;i<max_elements;i++)
-    {
-        if(ite != users.end())
-        {
-            dest_array[i] = *ite;
-            ite++;
-        }
-        else
-            dest_array[i] = 0;
-    }
-}
-
 QString makeCustomCommand(const QString& cmd, const QString& value)
 {
     return QString("%1\r\n%2").arg(cmd).arg(value);
@@ -162,129 +44,6 @@ QString makeCustomCommand(const QString& cmd, const QString& value)
 QStringList getCustomCommand(const MyTextMessage& msg)
 {
     return msg.moreMessage.split("\r\n");
-}
-
-void initDefaultAudioCodec(AudioCodec& audiocodec)
-{
-    audiocodec.nCodec = DEFAULT_AUDIOCODEC;
-    switch(DEFAULT_AUDIOCODEC)
-    {
-    case SPEEX_CODEC :
-        audiocodec.speex.nQuality = DEFAULT_SPEEX_QUALITY;
-        audiocodec.speex.nBandmode = DEFAULT_SPEEX_BANDMODE;
-        audiocodec.speex.nTxIntervalMSec = DEFAULT_SPEEX_DELAY;
-        audiocodec.speex.bStereoPlayback = DEFAULT_SPEEX_SIMSTEREO;
-        break;
-    case SPEEX_VBR_CODEC :
-        audiocodec.speex_vbr.nQuality = DEFAULT_SPEEX_VBR_QUALITY;
-        audiocodec.speex_vbr.nBandmode = DEFAULT_SPEEX_VBR_BANDMODE;
-        audiocodec.speex_vbr.nTxIntervalMSec = DEFAULT_SPEEX_VBR_DELAY;
-        audiocodec.speex_vbr.bStereoPlayback = DEFAULT_SPEEX_VBR_SIMSTEREO;
-        audiocodec.speex_vbr.nBitRate = DEFAULT_SPEEX_VBR_BITRATE;
-        audiocodec.speex_vbr.nMaxBitRate = DEFAULT_SPEEX_VBR_MAXBITRATE;
-        audiocodec.speex_vbr.bDTX = DEFAULT_SPEEX_VBR_DTX;
-        break;
-    case OPUS_CODEC :
-        audiocodec.opus.nApplication = DEFAULT_OPUS_APPLICATION;
-        audiocodec.opus.nSampleRate = DEFAULT_OPUS_SAMPLERATE;
-        audiocodec.opus.nChannels = DEFAULT_OPUS_CHANNELS;
-        audiocodec.opus.nTxIntervalMSec = DEFAULT_OPUS_DELAY;
-        audiocodec.opus.nComplexity = DEFAULT_OPUS_COMPLEXITY;
-        audiocodec.opus.bVBR = DEFAULT_OPUS_VBR;
-        audiocodec.opus.bVBRConstraint = DEFAULT_OPUS_VBRCONSTRAINT;
-        audiocodec.opus.bDTX = DEFAULT_OPUS_DTX;
-        audiocodec.opus.bFEC = DEFAULT_OPUS_FEC;
-        audiocodec.opus.nBitRate = DEFAULT_OPUS_BITRATE;
-        break;
-    default :
-        audiocodec.nCodec = NO_CODEC;
-        break;
-    }
-}
-
-AudioPreprocessor initDefaultAudioPreprocessor(AudioPreprocessorType preprocessortype)
-{
-    AudioPreprocessor preprocessor = {};
-    preprocessor.nPreprocessor = preprocessortype;
-    switch (preprocessor.nPreprocessor)
-    {
-    case NO_AUDIOPREPROCESSOR :
-        break;
-    case SPEEXDSP_AUDIOPREPROCESSOR :
-        preprocessor.speexdsp.bEnableAGC = DEFAULT_SPEEXDSP_AGC_ENABLE;
-        preprocessor.speexdsp.nGainLevel = DEFAULT_SPEEXDSP_AGC_GAINLEVEL;
-        preprocessor.speexdsp.nMaxIncDBSec = DEFAULT_SPEEXDSP_AGC_INC_MAXDB;
-        preprocessor.speexdsp.nMaxDecDBSec = DEFAULT_SPEEXDSP_AGC_DEC_MAXDB;
-        preprocessor.speexdsp.nMaxGainDB = DEFAULT_SPEEXDSP_AGC_GAINMAXDB;
-        preprocessor.speexdsp.bEnableDenoise = DEFAULT_SPEEXDSP_DENOISE_ENABLE;
-        preprocessor.speexdsp.nMaxNoiseSuppressDB = DEFAULT_SPEEXDSP_DENOISE_SUPPRESS;
-        preprocessor.speexdsp.bEnableEchoCancellation = DEFAULT_SPEEXDSP_ECHO_ENABLE;
-        preprocessor.speexdsp.nEchoSuppress = DEFAULT_SPEEXDSP_ECHO_SUPPRESS;
-        preprocessor.speexdsp.nEchoSuppressActive = DEFAULT_SPEEXDSP_ECHO_SUPPRESSACTIVE;
-        break;
-    case TEAMTALK_AUDIOPREPROCESSOR :
-        preprocessor.ttpreprocessor.nGainLevel = DEFAULT_TEAMTALK_GAINLEVEL;
-        preprocessor.ttpreprocessor.bMuteLeftSpeaker = DEFAULT_TEAMTALK_MUTELEFT;
-        preprocessor.ttpreprocessor.bMuteRightSpeaker = DEFAULT_TEAMTALK_MUTERIGHT;
-        break;
-    case WEBRTC_AUDIOPREPROCESSOR :
-        preprocessor.webrtc.preamplifier.bEnable = DEFAULT_WEBRTC_PREAMPLIFIER_ENABLE;
-        preprocessor.webrtc.preamplifier.fFixedGainFactor = DEFAULT_WEBRTC_PREAMPLIFIER_GAINFACTOR;
-        preprocessor.webrtc.levelestimation.bEnable = DEFAULT_WEBRTC_LEVELESTIMATION_ENABLE;
-        preprocessor.webrtc.voicedetection.bEnable = DEFAULT_WEBRTC_VAD_ENABLE;
-        preprocessor.webrtc.gaincontroller2.bEnable = DEFAULT_WEBRTC_GAINCTL_ENABLE;
-        preprocessor.webrtc.gaincontroller2.fixeddigital.fGainDB = DEFAULT_WEBRTC_GAINDB;
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.bEnable = DEFAULT_WEBRTC_SAT_PROT_ENABLE;
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fInitialSaturationMarginDB = DEFAULT_WEBRTC_INIT_SAT_MARGIN_DB;
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fExtraSaturationMarginDB = DEFAULT_WEBRTC_EXTRA_SAT_MARGIN_DB;
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fMaxGainChangeDBPerSecond = DEFAULT_WEBRTC_MAXGAIN_DBSEC;
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fMaxOutputNoiseLevelDBFS = DEFAULT_WEBRTC_MAX_OUT_NOISE;
-        preprocessor.webrtc.noisesuppression.bEnable = DEFAULT_WEBRTC_NOISESUPPRESS_ENABLE;
-        preprocessor.webrtc.noisesuppression.nLevel = DEFAULT_WEBRTC_NOISESUPPRESS_LEVEL;
-        preprocessor.webrtc.echocanceller.bEnable = DEFAULT_WEBRTC_ECHO_CANCEL_ENABLE;
-        break;
-    }
-    return preprocessor;
-}
-
-AudioPreprocessor loadAudioPreprocessor(AudioPreprocessorType preprocessortype)
-{
-    AudioPreprocessor preprocessor = initDefaultAudioPreprocessor(preprocessortype);
-    switch (preprocessor.nPreprocessor)
-    {
-    case NO_AUDIOPREPROCESSOR:
-        break;
-    case SPEEXDSP_AUDIOPREPROCESSOR:
-        preprocessor.speexdsp.bEnableAGC = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_ENABLE, DEFAULT_SPEEXDSP_AGC_ENABLE).toBool();
-        preprocessor.speexdsp.nGainLevel = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_GAINLEVEL, DEFAULT_SPEEXDSP_AGC_GAINLEVEL).toInt();
-        preprocessor.speexdsp.nMaxIncDBSec = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_INC_MAXDB, DEFAULT_SPEEXDSP_AGC_INC_MAXDB).toInt();
-        preprocessor.speexdsp.nMaxDecDBSec = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_DEC_MAXDB, DEFAULT_SPEEXDSP_AGC_DEC_MAXDB).toInt();
-        preprocessor.speexdsp.nMaxGainDB = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_GAINMAXDB, DEFAULT_SPEEXDSP_AGC_GAINMAXDB).toInt();
-        preprocessor.speexdsp.bEnableDenoise = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_DENOISE_ENABLE, DEFAULT_SPEEXDSP_DENOISE_ENABLE).toBool();
-        preprocessor.speexdsp.nMaxNoiseSuppressDB = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_DENOISE_SUPPRESS, DEFAULT_SPEEXDSP_DENOISE_SUPPRESS).toInt();
-        preprocessor.speexdsp.bEnableEchoCancellation = FALSE; // unusable for streaming
-        preprocessor.speexdsp.nEchoSuppress = DEFAULT_SPEEXDSP_ECHO_SUPPRESS;
-        preprocessor.speexdsp.nEchoSuppressActive = DEFAULT_SPEEXDSP_ECHO_SUPPRESSACTIVE;
-        break;
-    case TEAMTALK_AUDIOPREPROCESSOR:
-        preprocessor.ttpreprocessor.bMuteLeftSpeaker = ttSettings->value(SETTINGS_STREAMMEDIA_TTAP_MUTELEFT, false).toBool();
-        preprocessor.ttpreprocessor.bMuteRightSpeaker = ttSettings->value(SETTINGS_STREAMMEDIA_TTAP_MUTERIGHT, false).toBool();
-        preprocessor.ttpreprocessor.nGainLevel = ttSettings->value(SETTINGS_STREAMMEDIA_TTAP_GAINLEVEL, SOUND_GAIN_DEFAULT).toInt();
-        break;
-    case WEBRTC_AUDIOPREPROCESSOR :
-        preprocessor.webrtc.gaincontroller2.bEnable = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_GAINCTL_ENABLE, DEFAULT_WEBRTC_GAINCTL_ENABLE).toBool();
-        preprocessor.webrtc.gaincontroller2.fixeddigital.fGainDB = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_GAINDB, DEFAULT_WEBRTC_GAINDB).toFloat();
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.bEnable = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_SAT_PROT_ENABLE, DEFAULT_WEBRTC_SAT_PROT_ENABLE).toBool();
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fInitialSaturationMarginDB = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_INIT_SAT_MARGIN_DB, DEFAULT_WEBRTC_INIT_SAT_MARGIN_DB).toFloat();
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fExtraSaturationMarginDB = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_EXTRA_SAT_MARGIN_DB, DEFAULT_WEBRTC_EXTRA_SAT_MARGIN_DB).toFloat();
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fMaxGainChangeDBPerSecond = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_MAXGAIN_DBSEC, DEFAULT_WEBRTC_MAXGAIN_DBSEC).toFloat();
-        preprocessor.webrtc.gaincontroller2.adaptivedigital.fMaxOutputNoiseLevelDBFS = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_MAX_OUT_NOISE, DEFAULT_WEBRTC_MAX_OUT_NOISE).toFloat();
-        preprocessor.webrtc.noisesuppression.bEnable = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_NOISESUPPRESS_ENABLE, DEFAULT_WEBRTC_NOISESUPPRESS_ENABLE).toBool();
-        preprocessor.webrtc.noisesuppression.nLevel = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_NOISESUPPRESS_LEVEL, DEFAULT_WEBRTC_NOISESUPPRESS_LEVEL).toFloat();
-        preprocessor.webrtc.echocanceller.bEnable = FALSE; // unusable for streaming
-        break;
-    }
-    return preprocessor;
 }
 
 #if defined(Q_OS_WIN32)
@@ -341,21 +100,6 @@ bool isComputerIdle(int /*idle_secs*/)
     return false;
 }
 #endif
-
-bool isMyselfTalking()
-{
-    bool talking = false;
-    ClientFlags flags = TT_GetFlags(ttInst);
-    talking = (flags & CLIENT_TX_VOICE) ||
-              ((flags & CLIENT_SNDINPUT_VOICEACTIVATED) &&
-              (flags & CLIENT_SNDINPUT_VOICEACTIVE));
-    return talking;
-}
-
-bool isMyselfStreaming()
-{
-    return (TT_GetFlags(ttInst) & (CLIENT_STREAM_AUDIO | CLIENT_STREAM_VIDEO)) != CLIENT_CLOSED;
-}
 
 bool HostEntry::sameHost(const HostEntry& host, bool nickcheck) const
 {
@@ -603,189 +347,6 @@ void deleteServerEntry(const QString& name)
         setServerEntry(i, hosts[i]);
 }
 
-void processTrustedXML(const QDomElement& hostElement, HostEntry& entry)
-{
-    QDomElement trusted = hostElement.firstChildElement("trusted-certificate");
-    if (!trusted.isNull())
-    {
-        QDomElement tmp = trusted.firstChildElement("certificate-authority-pem");
-        if (!tmp.isNull())
-            entry.encryption.cacertdata = tmp.text();
-        tmp = trusted.firstChildElement("client-certificate-pem");
-        if (!tmp.isNull())
-            entry.encryption.certdata = tmp.text();
-        tmp = trusted.firstChildElement("client-private-key-pem");
-        if (!tmp.isNull())
-            entry.encryption.privkeydata = tmp.text();
-        tmp = trusted.firstChildElement("verify-peer");
-        if (!tmp.isNull())
-            entry.encryption.verifypeer = tmp.text() == "true";
-    }
-}
-
-void processAuthXML(const QDomElement& hostElement, HostEntry& entry)
-{
-    QDomElement auth = hostElement.firstChildElement("auth");
-    if (!auth.isNull())
-    {
-        QDomElement tmp = auth.firstChildElement("username");
-        if (!tmp.isNull())
-            entry.username = tmp.text();
-
-        tmp = auth.firstChildElement("password");
-        if (!tmp.isNull())
-            entry.password = tmp.text();
-
-        tmp = auth.firstChildElement("nickname");
-        if (!tmp.isNull())
-            entry.nickname = tmp.text();
-    }
-}
-
-void processJoinXML(const QDomElement& hostElement, HostEntry& entry)
-{
-    QDomElement join = hostElement.firstChildElement("join");
-    if(!join.isNull())
-    {
-        QDomElement tmp = join.firstChildElement("channel");
-        if (!tmp.isNull())
-            entry.channel = tmp.text();
-        tmp = join.firstChildElement("password");
-        if (!tmp.isNull())
-            entry.chanpasswd = tmp.text();
-    }
-}
-
-void processClientSetupXML(const QDomElement& hostElement, HostEntry& entry)
-{
-    QDomElement client = hostElement.firstChildElement(CLIENTSETUP_TAG);
-    if (!client.isNull())
-    {
-        QDomElement tmp = client.firstChildElement("nickname");
-        if (!tmp.isNull())
-            entry.nickname = tmp.text();
-
-        tmp = client.firstChildElement("gender");
-        if(!tmp.isNull())
-        {
-            switch (tmp.text().toInt())
-            {
-            case GENDER_MALE :
-                entry.gender = GENDER_MALE;
-                break;
-            case GENDER_FEMALE :
-                entry.gender = GENDER_FEMALE;
-                break;
-            case GENDER_NEUTRAL :
-            default:
-                entry.gender = GENDER_NEUTRAL;
-                break;
-            }
-        }
-#if defined(Q_OS_WIN32)
-        tmp = client.firstChildElement("win-hotkey");
-#elif defined(Q_OS_DARWIN)
-        tmp = client.firstChildElement("mac-hotkey");
-#elif defined(Q_OS_LINUX)
-        tmp = client.firstChildElement("x11-hotkey");
-#else
-#error Unknown OS
-#endif
-        if(!tmp.isNull())
-        {
-            tmp = tmp.firstChildElement("key");
-            while(!tmp.isNull())
-            {
-                entry.hotkey.push_back(tmp.text().toInt());
-                tmp = tmp.nextSiblingElement("key");
-            }
-        }
-        tmp = client.firstChildElement("voice-activated");
-        if(!tmp.isNull())
-            entry.voiceact = tmp.text().toInt();
-        tmp = client.firstChildElement("videoformat");
-        if(!tmp.isNull())
-        {
-            QDomElement cap = tmp.firstChildElement("width");
-            if(!cap.isNull())
-                entry.capformat.nWidth = cap.text().toInt();
-            cap = tmp.firstChildElement("height");
-            if(!cap.isNull())
-                entry.capformat.nHeight = cap.text().toInt();
-
-            cap = tmp.firstChildElement("fps-numerator");
-            if(!cap.isNull())
-                entry.capformat.nFPS_Numerator = cap.text().toInt();
-            cap = tmp.firstChildElement("fps-denominator");
-            if(!cap.isNull())
-                entry.capformat.nFPS_Denominator = cap.text().toInt();
-
-            cap = tmp.firstChildElement("fourcc");
-            if(!cap.isNull())
-                entry.capformat.picFourCC = (FourCC)cap.text().toInt();
-        }
-        tmp = client.firstChildElement("videocodec");
-        if(!tmp.isNull())
-        {
-            QDomElement vid = tmp.firstChildElement("codec");
-            if(!vid.isNull())
-                entry.vidcodec.nCodec = (Codec)vid.text().toInt();
-            switch(entry.vidcodec.nCodec)
-            {
-            case WEBM_VP8_CODEC :
-                vid = tmp.firstChildElement("webm-vp8-bitrate");
-                if(!vid.isNull())
-                {
-                    entry.vidcodec.webm_vp8.nRcTargetBitrate = vid.text().toInt();
-                    entry.vidcodec.webm_vp8.nEncodeDeadline = DEFAULT_WEBMVP8_DEADLINE;
-                }
-                break;
-            case SPEEX_CODEC :
-            case SPEEX_VBR_CODEC :
-            case OPUS_CODEC :
-            case NO_CODEC :
-                break;
-            }
-        }
-    }
-}
-
-bool getServerEntry(const QDomElement& hostElement, HostEntry& entry)
-{
-    Q_ASSERT(hostElement.tagName() == "host");
-    bool ok = true;
-    QDomElement tmp = hostElement.firstChildElement("name");
-    if(!tmp.isNull())
-        entry.name = tmp.text();
-    else ok = false;
-
-    tmp = hostElement.firstChildElement("address");
-    if(!tmp.isNull())
-        entry.ipaddr = tmp.text();
-    else ok = false;
-
-    tmp = hostElement.firstChildElement("tcpport");
-    if(!tmp.isNull())
-        entry.tcpport = tmp.text().toInt();
-    else ok = false;
-
-    tmp = hostElement.firstChildElement("udpport");
-    if(!tmp.isNull())
-        entry.udpport = tmp.text().toInt();
-    else ok = false;
-
-    tmp = hostElement.firstChildElement("encrypted");
-    if(!tmp.isNull())
-        entry.encrypted = (tmp.text().toLower() == "true" || tmp.text() == "1");
-
-    processTrustedXML(hostElement, entry);
-    processAuthXML(hostElement, entry);
-    processJoinXML(hostElement, entry);
-    processClientSetupXML(hostElement, entry);
-
-    return ok;
-}
-
 void addDesktopAccessEntry(const DesktopAccessEntry& entry)
 {
     QVector<DesktopAccessEntry> entries;
@@ -886,47 +447,44 @@ void deleteDesktopAccessEntries()
     }
 }
 
-QString parseXML(const QDomDocument& doc, QString elements)
+AudioPreprocessor loadAudioPreprocessor(AudioPreprocessorType preprocessortype)
 {
-    QDomElement element(doc.documentElement());
-    QStringList tokens = elements.split("/");
-    while (tokens.size())
+    AudioPreprocessor preprocessor = initDefaultAudioPreprocessor(preprocessortype);
+    switch (preprocessor.nPreprocessor)
     {
-        if (element.isNull() || element.nodeName() != tokens[0])
-            return QString();
-
-        tokens.removeFirst();
-        if (tokens.isEmpty())
-            return element.text();
-
-        element = element.firstChildElement(tokens[0]);
+    case NO_AUDIOPREPROCESSOR:
+        break;
+    case SPEEXDSP_AUDIOPREPROCESSOR:
+        preprocessor.speexdsp.bEnableAGC = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_ENABLE, DEFAULT_SPEEXDSP_AGC_ENABLE).toBool();
+        preprocessor.speexdsp.nGainLevel = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_GAINLEVEL, DEFAULT_SPEEXDSP_AGC_GAINLEVEL).toInt();
+        preprocessor.speexdsp.nMaxIncDBSec = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_INC_MAXDB, DEFAULT_SPEEXDSP_AGC_INC_MAXDB).toInt();
+        preprocessor.speexdsp.nMaxDecDBSec = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_DEC_MAXDB, DEFAULT_SPEEXDSP_AGC_DEC_MAXDB).toInt();
+        preprocessor.speexdsp.nMaxGainDB = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_AGC_GAINMAXDB, DEFAULT_SPEEXDSP_AGC_GAINMAXDB).toInt();
+        preprocessor.speexdsp.bEnableDenoise = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_DENOISE_ENABLE, DEFAULT_SPEEXDSP_DENOISE_ENABLE).toBool();
+        preprocessor.speexdsp.nMaxNoiseSuppressDB = ttSettings->value(SETTINGS_STREAMMEDIA_SPX_DENOISE_SUPPRESS, DEFAULT_SPEEXDSP_DENOISE_SUPPRESS).toInt();
+        preprocessor.speexdsp.bEnableEchoCancellation = FALSE; // unusable for streaming
+        preprocessor.speexdsp.nEchoSuppress = DEFAULT_SPEEXDSP_ECHO_SUPPRESS;
+        preprocessor.speexdsp.nEchoSuppressActive = DEFAULT_SPEEXDSP_ECHO_SUPPRESSACTIVE;
+        break;
+    case TEAMTALK_AUDIOPREPROCESSOR:
+        preprocessor.ttpreprocessor.bMuteLeftSpeaker = ttSettings->value(SETTINGS_STREAMMEDIA_TTAP_MUTELEFT, false).toBool();
+        preprocessor.ttpreprocessor.bMuteRightSpeaker = ttSettings->value(SETTINGS_STREAMMEDIA_TTAP_MUTERIGHT, false).toBool();
+        preprocessor.ttpreprocessor.nGainLevel = ttSettings->value(SETTINGS_STREAMMEDIA_TTAP_GAINLEVEL, SOUND_GAIN_DEFAULT).toInt();
+        break;
+    case WEBRTC_AUDIOPREPROCESSOR :
+        preprocessor.webrtc.gaincontroller2.bEnable = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_GAINCTL_ENABLE, DEFAULT_WEBRTC_GAINCTL_ENABLE).toBool();
+        preprocessor.webrtc.gaincontroller2.fixeddigital.fGainDB = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_GAINDB, DEFAULT_WEBRTC_GAINDB).toFloat();
+        preprocessor.webrtc.gaincontroller2.adaptivedigital.bEnable = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_SAT_PROT_ENABLE, DEFAULT_WEBRTC_SAT_PROT_ENABLE).toBool();
+        preprocessor.webrtc.gaincontroller2.adaptivedigital.fInitialSaturationMarginDB = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_INIT_SAT_MARGIN_DB, DEFAULT_WEBRTC_INIT_SAT_MARGIN_DB).toFloat();
+        preprocessor.webrtc.gaincontroller2.adaptivedigital.fExtraSaturationMarginDB = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_EXTRA_SAT_MARGIN_DB, DEFAULT_WEBRTC_EXTRA_SAT_MARGIN_DB).toFloat();
+        preprocessor.webrtc.gaincontroller2.adaptivedigital.fMaxGainChangeDBPerSecond = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_MAXGAIN_DBSEC, DEFAULT_WEBRTC_MAXGAIN_DBSEC).toFloat();
+        preprocessor.webrtc.gaincontroller2.adaptivedigital.fMaxOutputNoiseLevelDBFS = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_MAX_OUT_NOISE, DEFAULT_WEBRTC_MAX_OUT_NOISE).toFloat();
+        preprocessor.webrtc.noisesuppression.bEnable = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_NOISESUPPRESS_ENABLE, DEFAULT_WEBRTC_NOISESUPPRESS_ENABLE).toBool();
+        preprocessor.webrtc.noisesuppression.nLevel = ttSettings->value(SETTINGS_STREAMMEDIA_WEBRTC_NOISESUPPRESS_LEVEL, DEFAULT_WEBRTC_NOISESUPPRESS_LEVEL).toFloat();
+        preprocessor.webrtc.echocanceller.bEnable = FALSE; // unusable for streaming
+        break;
     }
-    return QString();
-}
-
-QString newVersionAvailable(const QDomDocument& updateDoc)
-{
-    return parseXML(updateDoc, "teamtalk/update/name");
-}
-
-QString downloadUpdateURL(const QDomDocument& updateDoc)
-{
-    return parseXML(updateDoc, "teamtalk/update/download-url");
-}
-
-QString newBetaVersionAvailable(const QDomDocument& updateDoc)
-{
-    return parseXML(updateDoc, "teamtalk/beta/name");
-}
-
-QString downloadBetaUpdateURL(const QDomDocument& updateDoc)
-{
-    return parseXML(updateDoc, "teamtalk/beta/download-url");
-}
-
-QString getBearWareRegistrationUrl(const QDomDocument& doc)
-{
-    return parseXML(doc, "teamtalk/bearware/register-url");
+    return preprocessor;
 }
 
 bool isWebLogin(const QString& username, bool includeParentLoginName)
@@ -987,189 +545,6 @@ void UserCached::sync(TTInstance* ttInst, const User& user)
     qDebug() << "Restored " << userCacheID(user);
 }
 
-
-QByteArray generateTTFile(const HostEntry& entry)
-{
-    QDomDocument doc(TTFILE_ROOT);
-    QDomProcessingInstruction pi = doc.createProcessingInstruction("xml", "version=\"1.0\" encoding=\"UTF-8\" ");
-    doc.appendChild(pi);
-
-    QDomElement root = doc.createElement(TTFILE_ROOT);
-    root.setAttribute("version", TTFILE_VERSION);
-    doc.appendChild(root);
-
-    QDomElement host = doc.createElement("host");
-
-    QDomElement name = doc.createElement("name");
-    name.appendChild(doc.createTextNode(entry.name));
-
-    QDomElement address = doc.createElement("address");
-    address.appendChild(doc.createTextNode(entry.ipaddr));
-
-    QDomElement tcpport = doc.createElement("tcpport");
-    tcpport.appendChild(doc.createTextNode(QString::number(entry.tcpport)));
-
-    QDomElement udpport = doc.createElement("udpport");
-    udpport.appendChild(doc.createTextNode(QString::number(entry.udpport)));
-
-    QDomElement encrypted = doc.createElement("encrypted");
-    encrypted.appendChild(doc.createTextNode(entry.encrypted?"true":"false"));
-
-    host.appendChild(name);
-    host.appendChild(address);
-    host.appendChild(tcpport);
-    host.appendChild(udpport);
-    host.appendChild(encrypted);
-
-    if (entry.encrypted)
-    {
-        QDomElement trusted = doc.createElement("trusted-certificate");
-        QDomElement ca_pem = doc.createElement("certificate-authority-pem");
-        ca_pem.appendChild(doc.createTextNode(entry.encryption.cacertdata));
-        QDomElement cert_pem = doc.createElement("client-certificate-pem");
-        cert_pem.appendChild(doc.createTextNode(entry.encryption.certdata));
-        QDomElement key_pem = doc.createElement("client-private-key-pem");
-        key_pem.appendChild(doc.createTextNode(entry.encryption.privkeydata));
-        QDomElement verifypeer = doc.createElement("verify-peer");
-        verifypeer.appendChild(doc.createTextNode(entry.encryption.verifypeer ? "true" : "false"));
-
-        trusted.appendChild(ca_pem);
-        trusted.appendChild(cert_pem);
-        trusted.appendChild(key_pem);
-        trusted.appendChild(verifypeer);
-        host.appendChild(trusted);
-    }
-
-    if(entry.username.size())
-    {
-        QDomElement auth = doc.createElement("auth");
-
-        QDomElement username = doc.createElement("username");
-        username.appendChild(doc.createTextNode(entry.username));
-
-        QDomElement password = doc.createElement("password");
-        password.appendChild(doc.createTextNode(entry.password));
-
-        QDomElement nickname = doc.createElement("nickname");
-        nickname.appendChild(doc.createTextNode(entry.nickname));
-
-        auth.appendChild(username);
-        auth.appendChild(password);
-        auth.appendChild(nickname);
-
-        host.appendChild(auth);
-    }
-
-    if(entry.channel.size())
-    {
-        QDomElement join = doc.createElement("join");
-
-        QDomElement channel = doc.createElement("channel");
-        channel.appendChild(doc.createTextNode(entry.channel));
-
-        QDomElement password = doc.createElement("password");
-        password.appendChild(doc.createTextNode(entry.chanpasswd));
-
-        join.appendChild(channel);
-        join.appendChild(password);
-
-        host.appendChild(join);
-    }
-
-    QDomElement client = doc.createElement(CLIENTSETUP_TAG);
-
-    if(entry.nickname.size())
-    {
-        QDomElement nickname = doc.createElement("nickname");
-        nickname.appendChild(doc.createTextNode(entry.nickname));
-        client.appendChild(nickname);
-    }
-    if(!entry.gender)
-    {
-        QDomElement gender = doc.createElement("gender");
-        gender.appendChild(doc.createTextNode(QString::number(entry.gender)));
-        client.appendChild(gender);
-    }
-    if(entry.hotkey.size())
-    {
-#if defined(Q_OS_WIN32)
-        QDomElement hotkey = doc.createElement("win-hotkey");
-#elif defined(Q_OS_DARWIN)
-        QDomElement hotkey = doc.createElement("mac-hotkey");
-#elif defined(Q_OS_LINUX)
-        QDomElement hotkey = doc.createElement("x11-hotkey");
-#else
-#error Unknown OS
-#endif
-        for (int k : entry.hotkey)
-        {
-            QDomElement key = doc.createElement("key");
-            key.appendChild(doc.createTextNode(QString::number(k)));
-            hotkey.appendChild(key);
-        }
-        client.appendChild(hotkey);
-    }
-    if(entry.voiceact >= 0)
-    {
-        QDomElement vox = doc.createElement("voice-activated");
-        vox.appendChild(doc.createTextNode(QString::number(entry.voiceact)));
-        client.appendChild(vox);
-    }
-    if(entry.capformat.nWidth)
-    {
-        QDomElement cap = doc.createElement("videoformat");
-
-        QDomElement newElement = doc.createElement("width");
-        newElement.appendChild(doc.createTextNode(QString::number(entry.capformat.nWidth)));
-        cap.appendChild(newElement);
-        newElement = doc.createElement("height");
-        newElement.appendChild(doc.createTextNode(QString::number(entry.capformat.nHeight)));
-        cap.appendChild(newElement);
-
-        newElement = doc.createElement("fps-numerator");
-        newElement.appendChild(doc.createTextNode(QString::number(entry.capformat.nFPS_Numerator)));
-        cap.appendChild(newElement);
-        newElement = doc.createElement("fps-denominator");
-        newElement.appendChild(doc.createTextNode(QString::number(entry.capformat.nFPS_Denominator)));
-        cap.appendChild(newElement);
-
-        newElement = doc.createElement("fourcc");
-        newElement.appendChild(doc.createTextNode(QString::number(entry.capformat.picFourCC)));
-        cap.appendChild(newElement);
-
-        client.appendChild(cap);
-    }
-    if(entry.vidcodec.nCodec != NO_CODEC)
-    {
-        QDomElement vidcodec = doc.createElement("videocodec");
-
-        QDomElement newElement = doc.createElement("codec");
-        newElement.appendChild(doc.createTextNode(QString::number(entry.vidcodec.nCodec)));
-        vidcodec.appendChild(newElement);
-        switch(entry.vidcodec.nCodec)
-        {
-        case WEBM_VP8_CODEC :
-            newElement = doc.createElement("webm-vp8-bitrate");
-            newElement.appendChild(doc.createTextNode(QString::number(entry.vidcodec.webm_vp8.nRcTargetBitrate)));
-            vidcodec.appendChild(newElement);
-            break;
-        case SPEEX_CODEC :
-        case SPEEX_VBR_CODEC :
-        case OPUS_CODEC :
-        case NO_CODEC :
-            break;
-        }
-        client.appendChild(vidcodec);
-    }
-
-    if(client.hasChildNodes())
-        host.appendChild(client);
-
-    root.appendChild(host);
-
-    return doc.toByteArray();
-}
-
 bool versionSameOrLater(const QString& check, const QString& against)
 {
     if(check == against) return true;
@@ -1204,7 +579,7 @@ QString getVersion(const User& user)
 
 QString limitText(const QString& text)
 {
-    int len = ttSettings->value(SETTINGS_DISPLAY_MAX_STRING, DEFAULT_MAX_STRING_LENGTH).toInt();
+    int len = ttSettings->value(SETTINGS_DISPLAY_MAX_STRING, SETTINGS_DISPLAY_MAX_STRING_DEFAULT).toInt();
     if(text.size()>len+3)
         return text.left(len) + "...";
     return text;
