+++ Client/qtTeamTalk/mainwindow.pri
@@ -22,7 +22,7 @@ HEADERS    = mainwindow.h preferencesdlg.h uservideowidget.h \
              statusbardlg.h statusbareventsmodel.h mycombobox.h \
              utilsound.h utilvideo.h utiltts.h utilui.h utilhotkey.h \
              serverlogeventsmodel.h mytreeview.h textmessagecontainer.h \
-             useraccountsmodel.h
+             useraccountsmodel.h encryptionsetupdlg.h utiltt.h utilxml.h
 
 SOURCES    = main.cpp mainwindow.cpp preferencesdlg.cpp uservideowidget.cpp \
              channelstree.cpp channeldlg.cpp userinfodlg.cpp \
@@ -41,7 +41,7 @@ SOURCES    = main.cpp mainwindow.cpp preferencesdlg.cpp uservideowidget.cpp \
              statusbardlg.cpp statusbareventsmodel.cpp mycombobox.cpp \
              utilsound.cpp utilvideo.cpp utiltts.cpp utilui.cpp utilhotkey.cpp \
              serverlogeventsmodel.cpp mytreeview.cpp textmessagecontainer.cpp \
-             useraccountsmodel.cpp
+             useraccountsmodel.cpp encryptionsetupdlg.cpp utiltt.cpp utilxml.cpp
 
 FORMS      = mainwindow.ui channel.ui preferences.ui \
              serverlist.ui userinfo.ui bannedusers.ui useraccounts.ui \
@@ -50,7 +50,7 @@ FORMS      = mainwindow.ui channel.ui preferences.ui \
              serverstats.ui onlineusers.ui mediastorage.ui userdesktop.ui \
              desktopshare.ui streammediafile.ui videotext.ui desktopaccess.ui \
              generatettfile.ui customvideofmt.ui bearwarelogindlg.ui \
-             audiopreprocessor.ui statusbardlg.ui
+             audiopreprocessor.ui statusbardlg.ui encryptionsetupdlg.ui
 
 RESOURCES += resources.qrc
